Added OepnGL 2D texture capabilities to the engine. Includes an example scene.

+++ OpenGL-Rendering-Engine/Exalted/src/Platform/OpenGL/OpenGLConfigurations.h
@@ -25,7 +25,7 @@ namespace Exalted
 		OpenGLConfigurations() = default;
 		~OpenGLConfigurations() = default;
 		static inline void SetPolygonMode(PolygonMode polygonMode) { glPolygonMode(GL_FRONT_AND_BACK, polygonMode); }
-		static inline void SetPointSize(unsigned int pointSize) { if(pointSize > 0) glPointSize(pointSize); }
+		static inline void SetPointSize(unsigned int pointSize) { if(pointSize > 0) glPointSize(static_cast<GLfloat>(pointSize)); }
 		static inline void EnableDepthTesting() { glEnable(GL_DEPTH_TEST); }
 		static inline void DisableDepthTesting() { glDisable(GL_DEPTH_TEST); }
 	};
